refactor: MAT-362 팀 매치 상세 패널 - 컴포넌트 폴더 구조 개선
