+++ src/routes/teams/$teamId/matches/-components/MatchDetailCard/GoalList/GoalList.tsx
