+++ src/routes/teams/$teamId/matches/-components/MatchDetailCard/GoalList/index.ts
