+++ src/routes/teams/$teamId/matches/-components/MatchDetailCard/MatchDetailLink/MatchDetailLink.css.ts
