+++ src/routes/teams/$teamId/matches/-components/MatchDetailCard/MatchMetadata/MatchMetadata.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { MatchMetadata } from './MatchMetadata';
 
 const meta = {
-  title: 'Routes/Teams/Matches/MatchDetailCard/Components/MatchMetadata',
+  title: 'Routes/Teams/Matches/MatchDetailCard/MatchMetadata',
   component: MatchMetadata,
   parameters: {
     layout: 'centered',
